Move logger to cartoframes.utils
