+++ cartoframes/__init__.py
@@ -1,7 +1,6 @@
 from ._version import __version__
 from .utils.utils import check_package
 from .core.cartodataframe import CartoDataFrame
-from .core.logger import set_log_level
 from .io.carto import read_carto, to_carto, has_table, delete_table, rename_table, \
                       copy_table, create_table_from_query, describe_table, update_privacy_table
 
@@ -23,6 +22,5 @@ __all__ = [
     'copy_table',
     'create_table_from_query',
     'describe_table',
-    'update_privacy_table',
-    'set_log_level'
+    'update_privacy_table'
 ]
