+++ cartoframes/core/managers/context_manager.py
@@ -8,7 +8,7 @@ from carto.exceptions import CartoException, CartoRateLimitException
 from carto.sql import SQLClient, BatchSQLClient, CopySQLClient
 
 from ... import __version__
-from ...core.logger import log
+from ...utils.logger import log
 from ...io.dataset_info import DatasetInfo
 from ...auth.defaults import get_default_credentials
 from ...utils.geom_utils import encode_geometry_ewkb
