+++ cartoframes/utils/__init__.py
@@ -1,0 +1,5 @@
+from .logger import set_log_level
+
+__all__ = [
+    'set_log_level'
+]
