+++ docs/includes/utils.rst
@@ -1,0 +1,8 @@
+Utils
+-----
+
+.. automodule:: cartoframes.utils
+    :members:
+    :member-order: bysource
+    :undoc-members:
+    :show-inheritance:
